refactor: consolidate tool usage parsing by removing redundant module and updating imports
